Refactor database table names
This commit refactors the database table names in the Prisma schema. The `Account` model is now mapped to the "account" table, the `Session` model is mapped to the "session" table, the `User` model is mapped to the "user" table, and the `VerificationToken` model is mapped to the "verification_token" table. Additionally, the `Game` model now has a `uuid` field instead of a `slug` field and is mapped to the "game" table.